flitering in tweetlist page remove from wall page
